+++ builders/CacheBuilder.d.ts
@@ -1,15 +1,18 @@
 export default class CacheBuilder {
     protected conf: Record<string, any>;
+    protected conn?: string;
     protected prefix: string;
     protected redis: Record<string, Function>;
     constructor();
     private get config();
+    private get currentConnection();
+    private get driver();
     private key;
-    private connection;
     private filePath;
     private file;
     private setFile;
     private getFile;
+    connection(conn: string): CacheBuilder;
     remember(key: string, callback: Function, ttl?: number): Promise<any>;
     has(key: string): Promise<boolean>;
     get(key: string): Promise<any>;