+++ enums/CacheDriverEnum.d.ts
@@ -1,0 +1,5 @@
+declare enum CacheDriverEnum {
+    Local = "local",
+    Redis = "redis"
+}
+export default CacheDriverEnum;