+++ enums/CacheDriverEnum.js
@@ -1,0 +1,6 @@
+var CacheDriverEnum;
+(function (CacheDriverEnum) {
+    CacheDriverEnum["Local"] = "local";
+    CacheDriverEnum["Redis"] = "redis";
+})(CacheDriverEnum || (CacheDriverEnum = {}));
+export default CacheDriverEnum;