+++ facades/Cache.d.ts
@@ -1,4 +1,6 @@
+import CacheBuilder from "../builders/CacheBuilder";
 export default class Cache {
+    static connection(connection: string): CacheBuilder;
     static remember(key: string, callback: Function, ttl?: number): Promise<any>;
     static has(key: string): Promise<boolean>;
     static get(key: string): Promise<any>;