+++ index.js
@@ -1,2 +1,3 @@
 export { default } from "./facades/Cache";
+export * from "./enums/index";
 export * from "./facades/index";