+++ src/config/cache.ts
@@ -1,14 +1,17 @@
 import App from "@bejibun/app";
+import CacheDriverEnum from "@/enums/CacheDriverEnum";
 
 const config: Record<string, any> = {
     connection: "local",
 
     connections: {
         local: {
+            driver: CacheDriverEnum.Local,
             path: App.Path.storagePath("cache") // absolute path
         },
 
         redis: {
+            driver: CacheDriverEnum.Redis,
             host: "127.0.0.1",
             port: 6379,
             password: "",