+++ src/enums/CacheDriverEnum.ts
@@ -1,0 +1,6 @@
+enum CacheDriverEnum {
+    Local = "local",
+    Redis = "redis"
+}
+
+export default CacheDriverEnum;