+++ src/index.ts
@@ -1,3 +1,5 @@
 export {default} from "@/facades/Cache";
 
+export * from "@/enums/index";
+
 export * from "@/facades/index";